+++ source/reg/menu.rb
@@ -51,8 +51,8 @@ module REG
           [
             100,
             TRANSLATE['Yes'],
-            -10,
-            10,
+            -10.0,
+            10.0,
             10.0
           ], # Defaults
           ['', TRANSLATE['Yes'] + '|' + TRANSLATE['No']], # List
@@ -64,8 +64,8 @@ module REG
 
         PARAMETERS[:entity_count] = parameters[0].to_i
         PARAMETERS[:rotate_entities?] = (parameters[1] == TRANSLATE['Yes'])
-        PARAMETERS[:entity_min_size] = parameters[2].to_i
-        PARAMETERS[:entity_max_size] = parameters[3].to_i
+        PARAMETERS[:entity_min_size] = parameters[2].to_f
+        PARAMETERS[:entity_max_size] = parameters[3].to_f
         PARAMETERS[:entity_density] = parameters[4].to_f
 
         Generator.new
